Migrate random tools from matlab_compat to random
