+++ src/aspire/source/xform.py
@@ -5,7 +5,7 @@ from joblib import Memory
 
 from aspire.image import Image
 from aspire.utils.filters import PowerFilter, ZeroFilter
-from aspire.utils.matlab_compat import randn
+from aspire.utils.random import randn
 
 logger = logging.getLogger(__name__)
