+++ src/aspire/utils/coor_trans.py
@@ -7,7 +7,7 @@ from numpy.linalg import norm
 from scipy.linalg import svd
 
 from aspire.utils import ensure
-from aspire.utils.matlab_compat import Random
+from aspire.utils.random import Random, randn
 
 
 def cart2pol(x, y):
