+++ tests/test_covar3d.py
@@ -12,7 +12,7 @@ from aspire.estimation.covar import CovarianceEstimator
 from aspire.estimation.mean import MeanEstimator
 from aspire.source.simulation import Simulation
 from aspire.utils.filters import RadialCTFFilter
-from aspire.utils.matlab_compat import Random
+from aspire.utils.random import Random
 from aspire.utils.matrix import eigs
 from aspire.utils.misc import src_wiener_coords
 from aspire.volume import Volume
