+++ tutorials/examples/cov3d_simulation.py
@@ -13,7 +13,7 @@ from aspire.estimation.mean import MeanEstimator
 from aspire.estimation.noise import WhiteNoiseEstimator
 from aspire.source.simulation import Simulation
 from aspire.utils.filters import RadialCTFFilter
-from aspire.utils.matlab_compat import Random
+from aspire.utils.random import Random
 from aspire.utils.matrix import eigs
 from aspire.utils.misc import src_wiener_coords
 from aspire.volume import Volume
